add attribute and update flag to only call getCount when the Table/Tablelike is modified.
repeated calls to getCount can slow down programs when used in loops, so recalculation should only be preformed if the object data is modified. This could be reverted back to the original implementation where the length is updated by the update functions, but leaving a function call to a base arc function is probably safer if for some reason the underlying data of the table was edited with something other than the active instance.